Add fsm to balancing subsystem

+++ mainboard/Src/peripherals/ltc6813_utils.c
@@ -210,8 +210,8 @@ void ltc6813_check_temperature(uint16_t temps, uint8_t index) {
 	}
 }
 
-void ltc6813_set_dcc(uint8_t indexes[], uint8_t cfgar[8], uint8_t cfgbr[8]) {
-	for (uint8_t i = 0; i < PACK_CELL_COUNT; i++) {
+void ltc6813_set_dcc(uint16_t indexes[], uint8_t cfgar[8], uint8_t cfgbr[8]) {
+	for (uint16_t i = 0; i < PACK_CELL_COUNT; i++) {
 		if (indexes[i] < 8) {
 			cfgar[4] += dcc[indexes[i]];
 		} else if (indexes[i] >= 8 && indexes[i] < 12) {
@@ -232,12 +232,12 @@ void ltc6813_set_dcc(uint8_t indexes[], uint8_t cfgar[8], uint8_t cfgbr[8]) {
 	}
 }
 
-void ltc6813_set_balancing(SPI_HandleTypeDef *hspi, uint8_t *indexes,
+void ltc6813_set_balancing(SPI_HandleTypeDef *hspi, uint16_t *indexes,
 						   int dcto) {
 	uint8_t cfgar[LTC6813_COUNT][8] = {0};
 	uint8_t cfgbr[LTC6813_COUNT][8] = {0};
 
-	for (uint8_t i = 0; i < LTC6813_COUNT; i++) {
+	for (uint16_t i = 0; i < LTC6813_COUNT; i++) {
 		// cfgbr[i][1] += 0b00001000; // set DTMEN
 		cfgar[i][0] += 0b00000010;	// set DTEN
 		cfgar[i][5] += dcto << 4;	// Set timer
